+++ Core.lua
@@ -32,7 +32,9 @@ function GearPolice:OnEnable()
     self:RegisterEvent("INSPECT_READY", "OnInspectReady")
     self:RegisterEvent("GROUP_ROSTER_UPDATE", "UpdateGroupMembers")
     self:RegisterEvent("PLAYER_REGEN_ENABLED", "OnCombatEnded")
+    self:RegisterEvent("PLAYER_TARGET_CHANGED", "OnPlayerTargetChanged")
     self:RegisterEvent("CHAT_MSG_WHISPER", "OnReportOfferWhisperReceived")
+    self:UpdatePlayerGearInfoWithGroupMembers()
     self:StartComms()
 end