+++ UI.lua
@@ -42,13 +42,6 @@ UI.ItemIconVisualStates = {
     },
 }
 
-UI.StatusTextures = {
-    scanning = "Interface\\COMMON\\Indicator-Yellow",
-    success = "Interface\\RaidFrame\\ReadyCheck-Ready",
-    failed = "Interface\\RaidFrame\\ReadyCheck-NotReady",
-    temporary_failed = "Interface\\RaidFrame\\ReadyCheck-Waiting",
-}
-
 UI.CheckStatusTextures = {
     InProgress = "Interface\\COMMON\\Indicator-Yellow",
     Successful = "Interface\\RaidFrame\\ReadyCheck-Ready",
@@ -56,10 +49,6 @@ UI.CheckStatusTextures = {
     Failed = "Interface\\RaidFrame\\ReadyCheck-NotReady",
     TemporaryFailed = "Interface\\RaidFrame\\ReadyCheck-Waiting",
 }
-
-function UI:GetStatusTexture(status)
-    return self.StatusTextures[status]
-end
 
 function UI:GetCheckStatusTexture(checkStatus)
     return self.CheckStatusTextures[checkStatus]